backend/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.3.2
# mediapipe>=0.10.8  # Optional - not available for Python 3.13 yet
# opencv-python>=4.8.1.78  # Optional - for AI models
# pytesseract>=0.3.10  # Optional - for OCR
# gtts>=2.5.1  # Optional - for TTS
# pydub>=0.25.1  # Optional - for audio processing
sqlalchemy>=2.0.23
aiosqlite>=0.19.0
alembic>=1.12.1
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.2

